Add example environment file for RTSP URLs 📡🎉 - Created a new `.env.example` file to provide templates for RTSP URL configurations. - This file includes placeholders for six RTSP URLs, guiding users on how to set up their environment variables.
The main reason for adding this is to facilitate easier setup for developers by giving a clear example of the required RTSP URL format. Now, when newbies dive into the codebase, they won't be left in the dark trying to figure out what the hell they need to put in their `.env` file. 🙌💻